requirements.txt
Jinja2==2.8
docker-py==1.6.0
docopt==0.6.2